pyproject.toml
[tool.poetry]
name = "music-platform-project"
version = "0.1.0"
description = ""
authors = ["abdelrahman ashour <abdelrahmanashour45@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
django-environ = "^0.11.2"
python-decouple = "^3.8"
django-imagekit = "^5.0.0"
pytest-django = "^4.8.0"
pytest = "^8.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
